requirements.txt
pytest-repeat
faker
playwright
pytest
pytest-xdist
pytest-playwright
python-dotenv
